+++ python/pathway/xpacks/llm/tests/test_vector_store.py
@@ -9,6 +9,7 @@ import pytest
 
 import pathway as pw
 from pathway.tests.utils import assert_table_equality
+from pathway.xpacks.llm import parsers
 from pathway.xpacks.llm.vector_store import VectorStoreServer
 
 
@@ -178,6 +179,46 @@ def test_async_embedder_preserves_params():
     _test_vs(fake_embeddings_model)
     _test_vs(fake_embeddings_model)
     assert call_count == 4  # dimension x 2 (no cache used), doc, query
+
+
+@pytest.mark.environment_changes  # unstructured parser adds env vars after first use
+@pytest.mark.parametrize("parser_cls", [parsers.ParseUnstructured])
+def test_vs_parsing(parser_cls):
+    def fake_embeddings_model(x: str) -> list[float]:
+        return [1.0, 1.0, 0.0]
+
+    docs = pw.debug.table_from_rows(
+        schema=pw.schema_from_types(data=bytes, _metadata=dict),
+        rows=[
+            (
+                "test".encode("utf-8"),
+                {"path": "pathway/xpacks/llm/tests/test_vector_store.py"},
+            )
+        ],
+    )
+
+    vector_server = VectorStoreServer(
+        docs,
+        parser=parser_cls(),
+        embedder=fake_embeddings_model,
+    )
+
+    retrieve_queries = pw.debug.table_from_markdown(
+        """
+        query | k | metadata_filter | filepath_globpattern
+        "Foo" | 1 |                 |
+        """,
+        schema=VectorStoreServer.RetrieveQuerySchema,
+    )
+
+    retrieve_outputs = vector_server.retrieve_query(retrieve_queries)
+    _, rows = pw.debug.table_to_dicts(retrieve_outputs)
+    (val,) = rows["result"].values()
+    assert isinstance(val, pw.Json)
+    (query_result,) = val.as_list()  # extract the single match
+    assert isinstance(query_result, dict)
+    assert query_result["dist"] < 1.0e-6  # type: ignore # the dist is not 0 due to float normalization
+    assert query_result["text"] == "test"
 
 
 @pytest.mark.parametrize(